+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/tags/PlaceObjectTag.java
@@ -286,4 +286,9 @@ public class PlaceObjectTag extends PlaceObjectTypeTag {
     public void setPlaceFlagHasMatrix(boolean placeFlagHasMatrix) {
 
     }
+
+    @Override
+    public boolean hasImage() {
+        return false;
+    }
 }