+++ CHANGELOG.md
@@ -20,6 +20,7 @@ All notable changes to this project will be documented in this file.
 - [#1647] Copying to clipboard - Transparency support
 - Incorrect placeobject display (tag selection)
 - Generic tag saving problem - timelined exception
+- [#1332] Flash viewer - Show directly added images when placeFlagHasImage is true on AS3 swfs
 
 ## [14.1.0] - 2021-03-05
 ### Added
@@ -2092,6 +2093,7 @@ All notable changes to this project will be documented in this file.
 [#1254]: https://www.free-decompiler.com/flash/issues/1254
 [#1636]: https://www.free-decompiler.com/flash/issues/1636
 [#1647]: https://www.free-decompiler.com/flash/issues/1647
+[#1332]: https://www.free-decompiler.com/flash/issues/1332
 [#1561]: https://www.free-decompiler.com/flash/issues/1561
 [#1623]: https://www.free-decompiler.com/flash/issues/1623
 [#1622]: https://www.free-decompiler.com/flash/issues/1622